requirements.txt
Django==3.2.19
django-generic-json-views~=0.8